requirements.txt
mysql-connector-python~=8.0.26
pyspark~=3.1.2